+++ src/engine/machine.ts
@@ -341,6 +341,7 @@ export class BotMachine {
         .some(pattern => {
           // extract message information
           const captures = execPattern(req.message, pattern);
+          // TODO: Remove `captures` in version 2.x, move to $flows.context
           const knowledges = {...req.variables, ...captures, $previous: req.previous, $input: req.message};
           this.logger.debug(`Explore dialogue for evaluation: ${pattern.source} => captures:`, captures);
 
@@ -362,6 +363,8 @@ export class BotMachine {
           req.currentDialogue = dialog.name;
           req.currentFlowIsResolved = true;
           req.variables = knowledges;
+          // assign session captured flows
+          Object.assign(req.$flows, captures);
 
           // add $ as the first matched variable for reply population
           if (captures.$1) {
