+++ test/engine/botscript.spec.ts
@@ -4,7 +4,6 @@ import { BotScript, Request } from '../../src/engine';
 describe('BotScript', () => {
 
   const bot = new BotScript();
-  let flowsRequest = new Request();
 
   bot.parse(`
   ! name BotScript
@@ -14,21 +13,6 @@ describe('BotScript', () => {
 
   + what is your name
   - My name is [name]
-
-  ~ age
-  - How old are you?
-  + I am #{age}
-  + #{age}
-
-  ~ email
-  - What is your email
-  + My email is *{email}
-
-  + my name is *{name}
-  + *{name} is my name
-  ~ age
-  ~ email
-  - Hello $name, you are $age and email $email!
 
   `);
 
@@ -43,41 +27,6 @@ describe('BotScript', () => {
       const req = new Request('what is your name');
       const res = await bot.handleAsync(req);
       assert.match(res.speechResponse, /my name is botscript/i, 'bot shows his name');
-    });
-  });
-
-  describe('basic dialogue flows', () => {
-    it('bot should ask human age', async () => {
-      const req = flowsRequest.enter('My name is Vu');
-      flowsRequest = await bot.handleAsync(req);
-      assert.isTrue(flowsRequest.isFlowing, 'enter dialogue flows!');
-      assert.match(flowsRequest.speechResponse, /how old are you/i, 'bot ask human\'s age');
-    });
-
-    it('bot should prompt again', async () => {
-      const req = flowsRequest.enter('something');
-      flowsRequest = await bot.handleAsync(req);
-      assert.isTrue(flowsRequest.isFlowing, 'still in dialogue flows!');
-      assert.match(flowsRequest.speechResponse, /how old are you/i, 'prompt one again');
-    });
-
-    it('bot should ask human email', async () => {
-      const req = flowsRequest.enter('20');
-      flowsRequest = await bot.handleAsync(req);
-      assert.isTrue(flowsRequest.isFlowing, 'still in dialogue flows!');
-      assert.equal(flowsRequest.variables.name, 'Vu', 'human name');
-      assert.equal(flowsRequest.variables.age, '20', 'human age');
-      assert.match(flowsRequest.speechResponse, /What is your email/i, 'bot send a next question');
-    });
-
-    it('bot should respond a greet with human name, age and email', async () => {
-      const req = flowsRequest.enter('my email is vunb@example.com');
-      flowsRequest = await bot.handleAsync(req);
-      assert.isFalse(flowsRequest.isFlowing, 'exit dialogue flows!');
-      assert.equal(flowsRequest.variables.name, 'Vu', 'human name');
-      assert.equal(flowsRequest.variables.age, '20', 'human age');
-      assert.equal(flowsRequest.variables.email, 'vunb@example.com', 'human email');
-      assert.match(flowsRequest.speechResponse, /hello/i, 'bot send a greeting');
     });
   });
